RedrawWindow: change flag type to uintptr

+++ user32.go
@@ -1032,7 +1032,7 @@ func KillTimer(hwnd HWND, nIDEvent uint32) bool {
 }
 
 // it will panic when the function fails
-func RedrawWindow(hWnd HWND, lpRect *RECT, hrgnUpdate HRGN, flag uint32) {
+func RedrawWindow(hWnd HWND, lpRect *RECT, hrgnUpdate HRGN, flag uintptr) {
 	ret, _, _ := procRedrawWindow.Call(
 		uintptr(hWnd),
 		uintptr(unsafe.Pointer(lpRect)),
